[APG-923] Refactor gender handling to use centralised Gender enum.
This commit moves the Gender enum to a dedicated package and replaces the String-based gender representation across the codebase. It ensures consistency, improves type safety, and reduces duplication by centralising gender logic in the shared `Gender` enum.
