+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/domain/entity/referencedata/type/Gender.kt
@@ -1,0 +1,7 @@
+package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.domain.entity.referencedata.type
+
+enum class Gender {
+  MALE,
+  FEMALE,
+  ANY,
+}
