+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsaccreditedprogrammesapi/restapi/model/CourseOffering.kt
@@ -2,6 +2,7 @@ package uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.restapi.model
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import io.swagger.v3.oas.annotations.media.Schema
+import uk.gov.justice.digital.hmpps.hmppsaccreditedprogrammesapi.domain.entity.referencedata.type.Gender
 import java.util.UUID
 
 /**
@@ -36,8 +37,3 @@ data class CourseOffering(
   @Schema(example = "M", description = "Gender for which course is offered")
   @get:JsonProperty("gender") val gender: Gender? = null,
 )
-
-enum class Gender {
-  MALE,
-  FEMALE,
-}
